Expand scale range to be at least DBL_EPSILON
A very small data range (from zero to a small floating point subnormal)
caused nTicks to be excessively big, ending up in a request for a huge
memory allocation for the axis.

This commit limits the number of ticks and the scale range to values
less than the floating point epsilon.